reorg function calling utils
